編集削除機能できたはず
コンフリクト直した

+++ src/views/TodoList.vue
@@ -7,23 +7,25 @@
       送信
     </button>
     <div v-for="(obj, index) in List" :key="index">
-      {{ obj.todo }} / {{ obj.limit }} :
-      <!--       <button class="delete" @click="show = !show">
-        ×
-      </button> -->
+      <input type="checkbox" v-model="obj.status" v-on:click="check(obj.id)" />
+      : {{ obj.todo }} / {{ obj.limit }} :
+      <button v-on:click="deletetodo(obj.id)">削除</button>
+      <button v-on:click="openModal(obj.id)">編集</button>
     </div>
-    <!--     <div v-show="show == false" id="overlay">
-      <div id="delateAlarm">
-        <p>この投稿を削除します</p>
-        <button @click="show = !show">
-          戻る
-        </button>
-        <button @click="deleteItem(index)">
-          削除
-        </button>
-      </div>
-    </div> -->
     <button v-on:click="logout">ログアウト</button>
+
+    <div id="overlay" v-show="showContent">
+      <div id="content">
+        <p>これがモーダルウィンドウです。</p>
+        todo :
+        <input type="text" v-model="text" />
+        期限 : <input type="date" v-model="timelimit" />
+        <button v-on:click="edittodo(editingId)">
+          送信
+        </button>
+        <button v-on:click="closeModal">Close</button>
+      </div>
+    </div>
   </div>
 </template>
 
@@ -41,10 +43,56 @@ export default {
       text: "",
       timelimit: "",
       now: "00:00:00",
-      show: true,
+      status: false,
+      showContent: false,
+      editingId: "",
     };
   },
   methods: {
+    openModal(index) {
+      this.editingId = index;
+      firebase
+        .firestore()
+        .collection("todo")
+        .doc(index)
+        .get()
+        .then(doc => {
+          this.text = doc.data().todo;
+          this.timelimit = doc.data().limit;
+        });
+      this.showContent = true;
+    },
+    closeModal: function() {
+      this.showContent = false;
+      this.List = [];
+      this.clearbox();
+      this.reload();
+    },
+    reload() {
+      this.List = [];
+      firebase
+        .firestore()
+        .collection("todo")
+        .where("user_id", "==", store.getters.getUserId)
+        .where("process_id", "==", store.getters.getProcessId)
+        .get()
+        .then(snapshot => {
+          snapshot.docs.forEach(doc => {
+            this.List.push({
+              id: doc.id,
+              ...doc.data(),
+            });
+          });
+          //ページ読み込んだと同時にvuexにもデータを同時に入れてる
+          const newTodos = this.List.map(todo => {
+            const obj = {};
+            obj.date = todo.limit;
+            obj.title = todo.todo;
+            return obj;
+          });
+          this.$store.dispatch("setTodoAction", { todos: newTodos });
+        });
+    },
     addlist() {
       var date = new Date();
       this.now =
@@ -59,7 +107,7 @@ export default {
         date.getMinutes() +
         ":" +
         date.getSeconds();
-      this.List.push(this.text);
+      /*       this.List.push(this.text); */
       firebase
         .firestore()
         .collection("todo")
@@ -67,9 +115,11 @@ export default {
           todo: this.text,
           created_at: this.now,
           limit: this.timelimit,
+          status: this.status,
           user_id: store.state.now_user_id,
           process_id: store.getters.getProcessId,
         });
+      this.List = [];
       firebase
         .firestore()
         .collection("todo")
@@ -92,21 +142,54 @@ export default {
         return obj;
       });
       this.$store.dispatch("setTodoAction", { todos: newTodos });
-      this.text == "";
-      this.timelimit == "";
+      this.clearbox();
     },
-    switchDelateAlarm() {
-      return false;
+    deletetodo(index) {
+      /*       console.log(index); */
+      var res = confirm("削除してもいいですか？");
+      if (res == true) {
+        firebase
+          .firestore()
+          .collection("todo")
+          .doc(index)
+          .delete()
+          .then(function() {
+            alert("削除しました");
+          })
+          .catch(function(error) {
+            alert.error("Error removing document: ", error);
+          });
+        this.reload();
+      }
     },
-    getIndex(index) {
-      this.delateId = this.List[index].id;
-    },
-    deleteItem(deleteId) {
+    edittodo(index) {
       firebase
+        .firestore()
         .collection("todo")
-        .doc(deleteId)
-        .delete();
-      this.show == false;
+        .doc(index)
+        .update({
+          todo: this.text,
+          limit: this.timelimit,
+        })
+        .then(() => {
+          this.reload();
+        });
+      this.closeModal();
+      this.clearbox();
+    },
+    clearbox() {
+      this.text = "";
+      this.timelimit = "";
+    },
+    check(index) {
+      firebase
+        .firestore()
+        .collection("todo")
+        .doc(index)
+        .get()
+        .add({
+          status: this.List.status,
+        });
     },
     logout() {
       firebase
@@ -148,3 +231,26 @@ export default {
   },
 };
 </script>
+
+<style>
+#overlay {
+  z-index: 1;
+
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.5);
+
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+#content {
+  z-index: 2;
+  width: 50%;
+  padding: 1em;
+  background: #fff;
+}
+</style>
